Corrects the setup process for docker (mimic-iii)
Minor corrections to the setup file for the creation of a containerized
instance of MIMIC-III.

- The owner of the mimiciii schema was set to "mimicuser". Such a user
does not exist. Changed to "mimic" since that is the name of the user
created in the previous portions of the script.

- The use of backticks around a code segment causes its evaluation in
bash. This was surely not meant to be done, so the backticks have been
escaped.